+++ src/app/[lang]/admin/crags/[cragSlug]/sectors/[sectorId]/routes/components/route-card/suggest-publish-route-dialog.tsx
@@ -2,7 +2,7 @@ import Dialog, { DialogSize } from "@/components/ui/dialog";
 import { Dispatch, SetStateAction, useState } from "react";
 import { useRouter } from "next/navigation";
 import { Route } from "@/graphql/generated";
-import updateRouteAction from "../server-actions/update-route-action";
+import updateRouteAction from "../../server-actions/update-route-action";
 
 type TSuggestPublishRouteDialogProps = {
   isOpen: boolean;
