+++ src/app/[lang]/admin/crags/[cragSlug]/sectors/[sectorId]/routes/page.tsx
@@ -123,6 +123,9 @@ gql`
         position
         created
         publishStatus
+        user {
+          fullName
+        }
       }
     }
   }
