+++ src/utils/text-helpers.ts
@@ -105,4 +105,10 @@ const pluralizeNoun = (
   }
 };
 
-export { pluralizeNoun };
+const genderizeVerb = (verb: string, gender: string) => {
+  if (gender == "F") return verb + "a";
+  if (gender == "M") return verb;
+  return verb + "/a";
+};
+
+export { pluralizeNoun, genderizeVerb };
